+++ lib/GraphDialog.js
@@ -1,4 +1,5 @@
 "use strict";
+var Parser_1 = require('./Parser');
 var Navigator_1 = require('./Navigator');
 var IntentScorer_1 = require('./IntentScorer');
 var interfaces = require('./Interfaces');
@@ -9,7 +10,8 @@ var NodeType = interfaces.NodeType;
 var GraphDialog = (function () {
     function GraphDialog(options) {
         this.options = options;
-        this.nav = new Navigator_1.Navigator(options);
+        var parser = new Parser_1.Parser(options.parser);
+        this.nav = new Navigator_1.Navigator(parser);
         this.intentScorer = new IntentScorer_1.IntentScorer();
         options.steps = options.steps || 100;
     }
@@ -56,7 +58,7 @@ var GraphDialog = (function () {
                 break;
             case NodeType.handler:
                 var handlerName = currentNode.data.name;
-                var handlerPath = path.join(this.options.handlersPath, handlerName);
+                var handlerPath = path.join(this.options.parser.handlersPath, handlerName);
                 var handler = require(handlerPath);
                 console.log('calling handler: ', currentNode.id, handlerName);
                 return handler(session, next, currentNode.data);
